requirements.txt
gymnasium==0.29.1
minigrid==3.0.0
numpy==2.2.6
PyYAML==6.0.2
torch==2.7.0
tqdm==4.67.1
